feat(executor): move wasm traps
