+++ crates/sandbox/Cargo.toml
@@ -13,6 +13,7 @@ readme = './README.md'
 keywords = ["wasm", "substrate", "runtime"]
 
 [dependencies]
+hex = "0.4.3"
 log = "0.4.14"
 bitflags = "1.2.1"
 serde = { version = "1", default-features = false, features = ["derive"] }
